create more usable tool scripts

+++ tools/coverage.sh
@@ -24,25 +24,52 @@
 
 set -e
 
-# delete all existing coverage data files and outputs
-find build-gcc/ -type f -name *.gcda -print0 | xargs -0 rm -f
+# handle arguments
+COMPILER=
+if [[ $# -gt 1 ]]; then
+    echo "Too many arguments"
+    exit 1
+fi
+if [[ $# -gt 0 ]]; then
+    case $1 in
+        --gcc)
+            COMPILER=gcc
+            shift
+            ;;
+        *)
+            echo "Unknown argument: $1"
+            exit 1
+            ;;
+    esac
+fi
 
-# generate initial coverage data
-lcov --capture --initial --base-directory . --directory build-gcc/src/ --output ./gadgets_init.lcov.info
+# set default options when not provided
+if [[ "$COMPILER" == "" ]]; then
+    COMPILER=gcc
+fi
 
-# execute the unit test runner
-pushd build-gcc/src/Tests
-./GadgetsTestRunner
-popd
+# run tests with coverage analysis (using gcov/lcov)
+if [[ "$COMPILER" == "gcc" ]]; then
+    # delete all existing coverage data files and outputs
+    find build-gcc/ -type f -name *.gcda -print0 | xargs -0 rm -f
 
-# generate coverage data after running tests
-lcov --capture --base-directory . --directory build-gcc/src/ --output ./gadgets_tests.lcov.info
+    # generate initial coverage data
+    lcov --capture --initial --base-directory . --directory build-gcc/src/ --output ./gadgets_init.lcov.info
 
-# combine the two sets of data, and then strip out the bits we don't need
-lcov --add ./gadgets_init.lcov.info --add ./gadgets_tests.lcov.info --output ./gadgets_all.lcov.info
-lcov --remove ./gadgets_all.lcov.info "/usr/*" "*/ucrt64/*" "*/mingw32/*" "*/mingw64/*" "*/clang32/*" "*/clang64/*" "`pwd`/lib/*" "`pwd`/src/Tests/*" "`pwd`/src/Example/*" --output ./gadgets_filtered.lcov.info
+    # execute the unit test runner
+    pushd build-gcc/src/Tests
+    ./GadgetsTestRunner
+    popd
 
-# generate coverage report
-rm -rf docs/coverage
-mkdir -p docs/coverage
-genhtml ./gadgets_filtered.lcov.info --prefix `pwd` --output-directory docs/coverage
+    # generate coverage data after running tests
+    lcov --capture --base-directory . --directory build-gcc/src/ --output ./gadgets_tests.lcov.info
+
+    # combine the two sets of data, and then strip out the bits we don't need
+    lcov --add ./gadgets_init.lcov.info --add ./gadgets_tests.lcov.info --output ./gadgets_all.lcov.info
+    lcov --remove ./gadgets_all.lcov.info "/usr/*" "*/ucrt64/*" "*/mingw32/*" "*/mingw64/*" "*/clang32/*" "*/clang64/*" "`pwd`/lib/*" "`pwd`/src/Tests/*" "`pwd`/src/Example/*" --output ./gadgets_filtered.lcov.info
+
+    # generate coverage report
+    rm -rf docs/coverage
+    mkdir -p docs/coverage
+    genhtml ./gadgets_filtered.lcov.info --prefix `pwd` --output-directory docs/coverage
+fi
